externals/obfuscation-analyzer/requirements.txt
networkx>=2.8
pyyaml>=6.0